requirements.txt
asgiref==3.7.2
Django==4.2.3
django-ckeditor==6.6.1
django-ckeditor-5==0.2.8
django-js-asset==2.1.0
django-taggit==4.0.0
Faker==19.1.0
Pillow==10.0.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
